(feat): Adding new quick-select dialog
This adds a new quick-select dialog much akin to the Command Palette,
except instead of selecting commands, it fuzzy finds on node name/path.

This somewhat improves the UX, in my opinion, for quickly switching
nodes. There is the `Ctrl+T` shortcut for "Find in Multiple Nodes...",
but that search box does not have suggestions or live filtering.

The inspiration for this feature comes from the `Ctrl+P` default binding
in VS Code where you can search for a file very quickly and switch to
it. I have put a default binding of `Ctrl+Alt+P` in, but that is a very
rough "what I could find that wasn't already used" kind of map and I am
very open to changing it.

Also note, if there is some reason this feature hasn't been in
CherryTree I am happy to abandon the PR. I just think it might make a
really cool addition.