Add cropping to insert image dialogue
Image widget is subclassed with draw and mouse events overridden.
Clicking and dragging selects an area, single clicking or right clicking
resets to full image.

Cropped out area is dimmed in the draw function.  There's then a
function to get the selected area and we use that to crop the Pixbuf
into the image the dialogue function returns.